PR-10 Change type from double to int

+++ app/src/main/java/com/pbs/expenseApp/database/entities/User.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class User(
     @PrimaryKey()
     val uuid: String,
-    val monthlySavings: Double
+    val monthlySavings: Int
 )
